Fix up the former novelonlinefree
