+++ README.md
@@ -206,7 +206,9 @@ The list of currently available sources and the future plans are given below:
 - [x] http://novelfull.com ![search](https://img.shields.io/badge/%F0%9F%94%8D-enabled-blue.svg)
 - [x] http://www.machinenoveltranslation.com
 - [x] http://zenithnovels.com
+- [x] https://4scanlation.xyz
 - [x] https://anythingnovel.com
+- [x] https://bestlightnovel.com (Formerly https://novelonlinefree.info) ![search](https://img.shields.io/badge/%F0%9F%94%8D-enabled-blue.svg)
 - [x] https://boxnovel.com ![search](https://img.shields.io/badge/%F0%9F%94%8D-enabled-blue.svg)
 - [x] https://comrademao.com
 - [x] https://creativenovels.com
@@ -219,7 +221,6 @@ The list of currently available sources and the future plans are given below:
 - [x] https://m.wuxiaworld.co ![search](https://img.shields.io/badge/%F0%9F%94%8D-enabled-blue.svg)
 - [x] https://meionovel.com
 - [x] https://mtled-novels.com ![search](https://img.shields.io/badge/%F0%9F%94%8D-enabled-blue.svg)
-- [x] https://novelonlinefree.info ![search](https://img.shields.io/badge/%F0%9F%94%8D-enabled-blue.svg)
 - [x] https://novelplanet.com ![search](https://img.shields.io/badge/%F0%9F%94%8D-enabled-blue.svg)
 - [x] https://volarenovels.com
 - [x] https://webnovel.online
@@ -240,7 +241,6 @@ The list of currently available sources and the future plans are given below:
 - [x] https://www.wuxiaworld.co ![search](https://img.shields.io/badge/%F0%9F%94%8D-enabled-blue.svg)
 - [x] https://www.wuxiaworld.com ![search](https://img.shields.io/badge/%F0%9F%94%8D-enabled-blue.svg)
 - [x] https://yukinovel.me
-- [x] https://4scanlation.xyz
 
 Rejected due to difficulty or other reasons:
