+++ README.pip
@@ -9,7 +9,9 @@ List of supported sites are given below.
 - http://novelfull.com
 - http://www.machinenoveltranslation.com
 - http://zenithnovels.com
+- https://4scanlation.xyz
 - https://anythingnovel.com
+- https://bestlightnovel.com
 - https://boxnovel.com
 - https://comrademao.com
 - https://creativenovels.com
@@ -22,7 +24,6 @@ List of supported sites are given below.
 - https://m.wuxiaworld.co
 - https://meionovel.com
 - https://mtled-novels.com
-- https://novelonlinefree.info
 - https://novelplanet.com
 - https://volarenovels.com
 - https://webnovel.online
@@ -42,6 +43,5 @@ List of supported sites are given below.
 - https://www.wuxiaworld.co
 - https://www.wuxiaworld.com
 - https://yukinovel.me
-- https://4scanlation.xyz
 
 Visit https://github.com/dipu-bd/lightnovel-crawler for more details.
